barbeque-nation-agent-main/Backend/requirements.txt
Flask
openai
google-api-python-client
google-auth
jinja2
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
